+++ CHANGELOG.md
@@ -57,6 +57,7 @@
 
 - Concatenate multilingualRoutes and externalRoutes (if available) to defaultRoutes @erral #3653
 - Fixed the `description` field not appearing in control panel fieldsets @JeffersonBledsoe #3696
+- Fixed "more" always show root contents @MdSahil-oss #3365
 - Add missing `--noninteractive` in the `build` script in package.json @sneridagh
 
 ### Internal